Adds .having function to fold, which realizes post-aggregation filters.
Summary:
This pull request adds the possibility of post-aggregation filtering in backtrace-forensics.

Here's how it works. Let's suppose we start with a query:
```
const query = BacktraceForensics.create().fold('a', 'head').fold('a', 'range').fold('a', 'distribution', 3);
```

Now we can add post-aggregation filters on folds that we have made in several ways:
```
// These do all the same
query.having('a', ['head'], 'equal', 123);
query.having('a', ['head'], '==', 123);
query.having('a', ['head'], '==', 0, 123); // 4th argument here specifies the index of value of fold result to filter upon
query.having('a', 0, '==', 123); // 2nd argument here specifies the index of folds being made on attribute 'a'
query.having('a', 0, '==', 0, 123);
```

For range, we could do it this way:
```
// These do all the same
query.having('a', ['range'], 'less-than', { from: 123, to: 456 });
query.having('a', ['range'], '<', { from: 123, to: 456 });
query.having('a', ['range'], '<', 0, 123).having('a', ['range'], '<', 1, 456);
query.having('a', 1, '<', { from: 123, to: 456 });
query.having('a', 1, '<', 0, 123).having('a', 1, '<', 1, 456);
```

And finally, for distribution:
```
query.having('a', ['distribution', 3], 'at-most', { keys: 123, tail: 456 });
query.having('a', 2, 'at-most', { keys: 123, tail: 456 });
// and so on
```

Post-aggregation filtering is enabled only for following folds: head, max, mean, min, sum, tail, unique, range (from, to), distribution (keys, tail).

I do not know if if is possible to filter on distribution/bin specific values. If it is possible, tell me how, and I'll include this in a future PR.

Test Plan: npm run test, manual

Reviewers: amadden, kdysput, rking, acronin, mrashkovsky, sfollari

Reviewed By: mrashkovsky

Differential Revision: https://hack.in.backtrace.io/D5911