Adds possibility of extending queries.
Summary:
This diff allows for extending `backtrace-forensics` within projects, or even to write plugins that extend functionality of queries.

See README diff for more info as how to construct them, and provided samples.

Exempt from README:

=Writing extensions=

As of version 0.3.0, `backtrace-forensics` supports writing extensions to its query functions.

To extend any of the queries, use one of the following functions:

-   `extendCoronerQuery` - use with `CommonCoronerQuery` interface
-   `extendSelectCoronerQuery` - use with `SelectCoronerQuery` interface
-   `extendSelectedCoronerQuery` - use with `SelectedCoronerQuery` interface
-   `extendFoldCoronerQuery` - use with `FoldCoronerQuery` interface
-   `extendFoldedCoronerQuery` - use with `FoldedCoronerQuery` interface

Be sure to add typings for Typescript.

You can check out sample plugins in the `samples` folder. The usage of these plugins is in the main sample in `extensions.ts`.

==Adding methods==

To add a method to a query, let's say fold queries:

```
// plugin.ts

extendFoldCoronerQuery({
    addHeadAndTailFolds(attribute: string) {
        return this.fold(attribute, 'head').fold(attribute, 'tail');
    },
});

extendFoldedCoronerQuery({
    addHourlyQuantizedColumn(attribute: string, name: string) {
        return this.virtualColumn(name, 'quantized_uint', {
            backing_column: attribute,
            size: 3600,
            offset: 86400,
        });
    },
});

declare module 'backtrace-forensics' {
    interface FoldCoronerQuery {
        // For not folded queries, be sure to return FoldedCoronerQuery to continue folding
        addHeadAndTailFolds(attribute: string): FoldedCoronerQuery;
    }

    interface FoldedCoronerQuery {
        // For already-folded queries, return FoldedCoronerQuery, or just this
        addHourlyQuantizedColumn(attribute: string, name: string): this;
    }
}
```

Now, as long as you have this module imported, you can use this method:

```
import './plugin';

BacktraceForensics.create()
    .filter('a', 'equals', 'b')
    .addHeadAndTailFolds('a')
    .fold('c', 'unique')
    .addHourlyQuantizedColumn('x', 'timestamp');
```

Test Plan:
check if the correct queries are being outputted
check if types are correct

Reviewers: kdysput, amadden, mrashkovsky, sfollari, acronin, rking

Reviewed By: kdysput, sfollari, rking

Differential Revision: https://hack.in.backtrace.io/D7234